backend/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
mistralai==0.1.0